+++ src/oic/oauth2/message.py
@@ -2,6 +2,7 @@ import copy
 import json
 import logging
 from collections import MutableMapping
+from collections import namedtuple
 from urllib.parse import parse_qs
 from urllib.parse import urlencode
 
@@ -151,6 +152,18 @@ class Message(MutableMapping):
         for key, val in self.c_default.items():
             self._dict[key] = val
 
+    @staticmethod
+    def _extract_cparam(key, _spec):
+        """Extract ParamDefinition for a given key.
+
+        The key can be direct attribute or lang typed attribute.
+        If ParamDefinition is not found, tries to return "*" attribute, if it exists, otherwise returns None.
+        """
+        for _key in (key, key.split('#')[0], '*'):
+            if _key in _spec:
+                return _spec[_key]
+        return None
+
     def to_urlencoded(self, lev=0):
         """
         Creates a string using the application/x-www-form-urlencoded format
@@ -160,26 +173,21 @@ class Message(MutableMapping):
 
         _spec = self.c_param
         if not self.lax:
-            for attribute, (_, req, _ser, _, _) in _spec.items():
-                if req and attribute not in self._dict:
+            for attribute, cparam in _spec.items():
+                if cparam.required and attribute not in self._dict:
                     raise MissingRequiredAttribute("%s" % attribute,
                                                    "%s" % self)
 
         params = []
 
         for key, val in self._dict.items():
-            try:
-                (_, req, _ser, _, null_allowed) = _spec[key]
-            except KeyError:  # extra attribute
-                try:
-                    _key, lang = key.split("#")
-                    (_, req, _ser, _deser, null_allowed) = _spec[_key]
-                except (ValueError, KeyError):
-                    try:
-                        (_, req, _ser, _, null_allowed) = _spec['*']
-                    except KeyError:
-                        _ser = None
-                        null_allowed = False
+            cparam = self._extract_cparam(key, _spec)
+            if cparam is not None:
+                _ser = cparam.serializer
+                null_allowed = cparam.null_allowed
+            else:
+                _ser = None
+                null_allowed = False
 
             if val is None and null_allowed is False:
                 continue
@@ -250,36 +258,28 @@ class Message(MutableMapping):
         _spec = self.c_param
 
         for key, val in parse_qs(urlencoded).items():
-            try:
-                (typ, _, _, _deser, _) = _spec[key]
-            except KeyError:
-                try:
-                    _key, lang = key.split("#")
-                    (typ, _, _, _deser, null_allowed) = _spec[_key]
-                except (ValueError, KeyError):
-                    try:
-                        (typ, _, _, _deser, null_allowed) = _spec['*']
-                    except KeyError:
-                        if len(val) == 1:
-                            val = val[0]
+            cparam = self._extract_cparam(key, _spec)
+            if cparam is None:
+                if len(val) == 1:
+                    val = val[0]
 
-                        self._dict[key] = val
-                        continue
+                self._dict[key] = val
+                continue
 
-            if isinstance(typ, list):
-                if _deser:
-                    self._dict[key] = _deser(val[0], "urlencoded")
+            if isinstance(cparam.type, list):
+                if cparam.deserializer is not None:
+                    self._dict[key] = cparam.deserializer(val[0], "urlencoded")
                 else:
                     self._dict[key] = val
             else:  # must be single value
                 if len(val) == 1:
-                    if _deser:
-                        self._dict[key] = _deser(val[0], "urlencoded")
-                    elif isinstance(val[0], typ):
+                    if cparam.deserializer is not None:
+                        self._dict[key] = cparam.deserializer(val[0], "urlencoded")
+                    elif isinstance(val[0], cparam.type):
                         self._dict[key] = val[0]
                     else:
                         try:
-                            self._dict[key] = typ(val[0])
+                            self._dict[key] = cparam.type(val[0])
                         except KeyError:
                             raise ParameterError(key)
                 else:
@@ -299,18 +299,11 @@ class Message(MutableMapping):
         _res = {}
         lev += 1
         for key, val in self._dict.items():
-            try:
-                (_, _, _ser, _, null_allowed) = _spec[str(key)]
-            except KeyError:
-                try:
-                    _key, _ = key.split("#")
-                    (_, req, _ser, _, null_allowed) = _spec[_key]
-                except (ValueError, KeyError):
-                    try:
-                        (_, req, _ser, _, null_allowed) = _spec['*']
-                    except KeyError:
-                        _ser = None
-
+            cparam = self._extract_cparam(key, _spec)
+            if cparam is not None:
+                _ser = cparam.serializer
+            else:
+                _ser = None
             if _ser:
                 val = _ser(val, "dict", lev)
 
@@ -335,25 +328,11 @@ class Message(MutableMapping):
         _spec = self.c_param
 
         for key, val in dictionary.items():
-            # Earlier versions of python don't like unicode strings as
-            # variable names
-            if val == "" or val == [""]:
+            if val in ('', ['']):
                 continue
-
-            skey = str(key)
-            if key in _spec:
-                lookup_key = key
-            else:
-                # might be a parameter with a lang tag
-                try:
-                    _key, _ = skey.split("#")
-                except ValueError:
-                    lookup_key = '*'
-                else:
-                    lookup_key = _key
-            if lookup_key in _spec:
-                (vtyp, _, _, _deser, null_allowed) = _spec[lookup_key]
-                self._add_value(skey, vtyp, key, val, _deser, null_allowed)
+            cparam = self._extract_cparam(key, _spec)
+            if cparam is not None:
+                self._add_value(key, cparam.type, key, val, cparam.deserializer, cparam.null_allowed)
             else:
                 self._dict[key] = val
         return self
@@ -711,40 +690,34 @@ class Message(MutableMapping):
         except KeyError:
             _allowed = {}
 
-        for (attribute, (typ, required, _, _, na)) in _spec.items():
+        for attribute, cparam in _spec.items():
             if attribute == "*":
                 continue
 
-            try:
-                val = self._dict[attribute]
-            except KeyError:
-                if required:
+            val = self._dict.get(attribute)
+            if val is None:
+                if cparam.required:
                     raise MissingRequiredAttribute("%s" % attribute)
                 continue
-            else:
-                if typ == bool:
-                    pass
-                elif not val:
-                    if required:
-                        raise MissingRequiredAttribute("%s" % attribute)
-                    continue
+            if cparam.type != bool and not val:
+                if cparam.required:
+                    raise MissingRequiredAttribute("%s" % attribute)
+                continue
 
             if attribute not in _allowed:
                 continue
 
-            if isinstance(typ, tuple):
-                _ityp = None
-                for _typ in typ:
+            if isinstance(cparam.type, tuple):
+                for _typ in cparam.type:
                     try:
                         self._type_check(_typ, _allowed[attribute], val)
-                        _ityp = _typ
                         break
                     except ValueError:
                         pass
-                if _ityp is None:
-                    raise NotAllowedValue(val)
+                    else:
+                        raise NotAllowedValue(val)
             else:
-                self._type_check(typ, _allowed[attribute], val, na)
+                self._type_check(cparam.type, _allowed[attribute], val, cparam.null_allowed)
 
         return True
 
@@ -788,8 +761,8 @@ class Message(MutableMapping):
 
     def __setitem__(self, key, value):
         try:
-            (vtyp, _, _, _deser, na) = self.c_param[key]
-            self._add_value(str(key), vtyp, key, value, _deser, na)
+            cparam = self.c_param[key]
+            self._add_value(str(key), cparam.type, key, value, cparam.deserializer, cparam.null_allowed)
         except KeyError:
             self._dict[key] = value
 
@@ -952,19 +925,15 @@ VSER = 2
 VDESER = 3
 VNULLALLOWED = 4
 
-SINGLE_REQUIRED_STRING = (str, True, None, None, False)
-SINGLE_OPTIONAL_STRING = (str, False, None, None, False)
-SINGLE_OPTIONAL_INT = (int, False, None, None, False)
-OPTIONAL_LIST_OF_STRINGS = ([str], False, list_serializer,
-                            list_deserializer, False)
-REQUIRED_LIST_OF_STRINGS = ([str], True, list_serializer,
-                            list_deserializer, False)
-OPTIONAL_LIST_OF_SP_SEP_STRINGS = ([str], False, sp_sep_list_serializer,
-                                   sp_sep_list_deserializer, False)
-REQUIRED_LIST_OF_SP_SEP_STRINGS = ([str], True, sp_sep_list_serializer,
-                                   sp_sep_list_deserializer, False)
-SINGLE_OPTIONAL_JSON = (str, False, json_serializer, json_deserializer,
-                        False)
+ParamDefinition = namedtuple('ParamDefinition', ['type', 'required', 'serializer', 'deserializer', 'null_allowed'])
+SINGLE_REQUIRED_STRING = ParamDefinition(str, True, None, None, False)
+SINGLE_OPTIONAL_STRING = ParamDefinition(str, False, None, None, False)
+SINGLE_OPTIONAL_INT = ParamDefinition(int, False, None, None, False)
+OPTIONAL_LIST_OF_STRINGS = ParamDefinition([str], False, list_serializer, list_deserializer, False)
+REQUIRED_LIST_OF_STRINGS = ParamDefinition([str], True, list_serializer, list_deserializer, False)
+OPTIONAL_LIST_OF_SP_SEP_STRINGS = ParamDefinition([str], False, sp_sep_list_serializer, sp_sep_list_deserializer, False)
+REQUIRED_LIST_OF_SP_SEP_STRINGS = ParamDefinition([str], True, sp_sep_list_serializer, sp_sep_list_deserializer, False)
+SINGLE_OPTIONAL_JSON = ParamDefinition(str, False, json_serializer, json_deserializer, False)
 
 REQUIRED = [SINGLE_REQUIRED_STRING, REQUIRED_LIST_OF_STRINGS,
             REQUIRED_LIST_OF_SP_SEP_STRINGS]